iOS 연동 코드상으로 95% 쯤 구현. !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!xcode 끄고 받아야됨.!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!

+++ Classes/AccountManagerPopup.cpp
@@ -571,7 +571,7 @@ bool AccountManagerPopup::init(int touchP)
 	{
 		facebookLogin->setFunction([=](CCObject*){
 #if CC_TARGET_PLATFORM == CC_PLATFORM_ANDROID
-			tryLogin(HSPMapping::kFACEBOOK, "Facebook ID", HSPLogin::FACEBOOK);
+			tryLogin((int)HSPMapping::kFACEBOOK, "Facebook ID", HSPLogin::FACEBOOK);
 #elif CC_TARGET_PLATFORM == CC_PLATFORM_IOS
 			tryLogin(HSPMappingType::HSP_MAPPINGTYPE_FACEBOOK, "Facebook ID", HSPLogin::FACEBOOK);
 #endif
